requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.35
asyncpg==0.30.0
python-dotenv==1.0.1
APScheduler==3.10.4
openai==1.54.0
